# packages/types/src/graphic/resources/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Canvas Render API
  description: todo
  contact:
    name: dyn.art
    url: https://dyn.art/?source=apidocs
  license:
    name: AGPL-3.0-or-later
  version: 1.0.0
paths:
...
  /v1/cnv/render:
    post:
      tags:
      - crate::routes::v1::cnv::render
      operationId: post_v1_cnv_render_handler
      parameters:
      - name: format
        in: path
        required: true
        schema:
          allOf:
          - $ref: '#/components/schemas/FileFormat'
          nullable: true
      - name: scriptArgs
        in: path
        required: true
        schema:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/LuaScriptArgsMap'
          nullable: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtifCanvas'
        required: true
      responses:
        '200':
          description: Generation success
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
...
components:
  schemas:
...
    AppError:
      type: object
      required:
      - status
      - code
      properties:
        additionalErrors:
          type: array
          items:
            type: object
            additionalProperties: {}
          description: An array of additional error details or nested errors that occurred during the process.
        code:
          $ref: '#/components/schemas/ErrorCode'
        description:
          type: string
          description: |-
            A detailed, human-readable description of the error. Provides additional context
            and, if applicable, steps to resolve the issue.
          nullable: true
        status:
          type: integer
          format: int32
          description: HTTP status code associated with the error, indicating the nature of the failure.
          minimum: 0
        uri:
          type: string
          description: An optional URI linking to a document or resource with more information about the error.
          nullable: true
...
    AssetContent:
      oneOf:
      - type: object
        description: Content stored as binary data.
        required:
        - content
        - type
        properties:
          content:
            type: string
            format: binary
          type:
            type: string
            enum:
            - Binary
      - type: object
        description: Content referenced by a URL.
        required:
        - url
        - type
        properties:
          type:
            type: string
            enum:
            - Url
          url:
            type: string
      discriminator:
        propertyName: type
    AssetContentType:
      oneOf:
      - type: object
        required:
        - type
        properties:
          type:
            type: string
            enum:
            - Unknown
      - type: object
        required:
        - type
        properties:
          type:
            type: string
            enum:
            - Jpeg
      - type: object
        required:
        - type
        properties:
          type:
            type: string
            enum:
            - Png
      - type: object
        required:
        - width
        - height
        - type
        properties:
          height:
            type: integer
            format: int32
            minimum: 0
          type:
            type: string
            enum:
            - Svg
          width:
            type: integer
            format: int32
            minimum: 0
      - type: object
        required:
        - type
        properties:
          type:
            type: string
            enum:
            - Ttf
      discriminator:
        propertyName: type
    AssetWithId:
      type: object
      required:
      - content
      - contentType
      properties:
        content:
          $ref: '#/components/schemas/AssetContent'
        contentType:
          $ref: '#/components/schemas/AssetContentType'
        id:
          allOf:
          - $ref: '#/components/schemas/ReferenceId'
          nullable: true
...
    CnvVersion:
      type: string
      enum:
      - V000001
...
    CoreInputEvent:
      oneOf:
      - allOf:
        - $ref: '#/components/schemas/UpdateCanvasSizeInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateCanvasSize
      - allOf:
        - $ref: '#/components/schemas/UpdateCanvasViewportInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateCanvasViewport
      - allOf:
        - $ref: '#/components/schemas/FocusRootNodesInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - FocusRootNodes
      - allOf:
        - $ref: '#/components/schemas/CreateNodeInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - CreateNode
      - allOf:
        - $ref: '#/components/schemas/UpdateFrameNodeInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateFrameNode
      - allOf:
        - $ref: '#/components/schemas/UpdateEllipseNodeInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateEllipseNode
      - allOf:
        - $ref: '#/components/schemas/UpdateStarNodeInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateStarNode
      - allOf:
        - $ref: '#/components/schemas/UpdatePolygonNodeInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdatePolygonNode
      - allOf:
        - $ref: '#/components/schemas/UpdateTextNodeInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateTextNode
      - allOf:
        - $ref: '#/components/schemas/UpdateFillStyleInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateFillStyle
      - allOf:
        - $ref: '#/components/schemas/UpdateStorkeStyleInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateStrokeStyle
      - allOf:
        - $ref: '#/components/schemas/UpdateDropShadowStyleInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateDropShadowStyle
      - allOf:
        - $ref: '#/components/schemas/CreatePaintInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - CreatePaint
      - allOf:
        - $ref: '#/components/schemas/UpdateSolidPaintInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateSolidPaint
      - allOf:
        - $ref: '#/components/schemas/UpdateImagePaintInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateImagePaint
      - allOf:
        - $ref: '#/components/schemas/UpdateGradientPaintInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateGradientPaint
      - allOf:
        - $ref: '#/components/schemas/CreateAssetInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - CreateAsset
      - allOf:
        - $ref: '#/components/schemas/DeleteEntityInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - DeleteEntity
      - allOf:
        - $ref: '#/components/schemas/UpdateEntityTransformInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateEntityTransform
      - allOf:
        - $ref: '#/components/schemas/UpdateEntitySizeInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateEntitySize
      - allOf:
        - $ref: '#/components/schemas/MoveEntityInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - MoveEntity
      - allOf:
        - $ref: '#/components/schemas/UpdateEntityRotationInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateEntityRotation
      - allOf:
        - $ref: '#/components/schemas/UpdateEntityVisibilityInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateEntityVisibility
      - allOf:
        - $ref: '#/components/schemas/UpdateEntityCornerRadiiInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateEntityCornerRadii
      - allOf:
        - $ref: '#/components/schemas/UpdateEntityBlendModeInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateEntityBlendMode
      - allOf:
        - $ref: '#/components/schemas/UpdateEntityOpacityInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateEntityOpacity
      - allOf:
        - $ref: '#/components/schemas/UpdateEntityChildrenInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - UpdateEntityChildren
      - allOf:
        - $ref: '#/components/schemas/RegisterLuaScriptInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - RegisterLuaScript
      - allOf:
        - $ref: '#/components/schemas/ExecuteLuaScriptInputEvent'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - ExecuteLuaScript
      discriminator:
        propertyName: type
...
    CreateAssetInputEvent:
      type: object
      required:
      - asset
      properties:
        asset:
          $ref: '#/components/schemas/AssetWithId'
    CreateNodeInputEvent:
      type: object
      required:
      - node
      properties:
        node:
          $ref: '#/components/schemas/Node'
    CreatePaintInputEvent:
      type: object
      required:
      - paint
      properties:
        paint:
          $ref: '#/components/schemas/Paint'
    DeleteEntityInputEvent:
      type: object
      required:
      - id
      properties:
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
...
    DtifCanvas:
      type: object
      description: |-
        DTIF (Design Tree Interchange Format) utilizes a flat structure for easy readability
        and efficient access & manipulation of design elements (Nodes, Paints, ..).
        https://softwareengineering.stackexchange.com/questions/350623/flat-or-nested-json-for-hierarchal-data
      required:
      - size
      - nodes
      properties:
        assets:
          type: array
          items:
            $ref: '#/components/schemas/AssetWithId'
          description: A list of assets.
        events:
          type: array
          items:
            $ref: '#/components/schemas/CoreInputEvent'
          description: A list of input events.
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/Node'
          description: A list of nodes.
        paints:
          type: array
          items:
            $ref: '#/components/schemas/Paint'
          description: A list of paints.
        scripts:
          type: array
          items:
            $ref: '#/components/schemas/LuaScriptWithId'
          description: A list of scripts.
        size:
          $ref: '#/components/schemas/Size'
        version:
          allOf:
          - $ref: '#/components/schemas/CnvVersion'
          nullable: true
        viewport:
          allOf:
          - $ref: '#/components/schemas/Viewport'
          nullable: true
    EllipseNode:
      type: object
      required:
      - size
      properties:
        blendMode:
          $ref: '#/components/schemas/BlendMode'
        endingAngle:
          type: number
          format: float
        id:
          allOf:
          - $ref: '#/components/schemas/ReferenceId'
          nullable: true
        innerRadiusRatio:
          type: number
          format: float
        layoutElement:
          $ref: '#/components/schemas/LayoutElement'
        opacity:
          $ref: '#/components/schemas/Opacity'
        rotationDeg:
          $ref: '#/components/schemas/Angle'
        size:
          $ref: '#/components/schemas/Size'
        startingAngle:
          type: number
          format: float
        styles:
          type: array
          items:
            $ref: '#/components/schemas/Style'
        translation:
          $ref: '#/components/schemas/Vec2'
        visible:
          type: boolean
...
    ErrorCode:
      type: string
    ExecuteLuaScriptInputEvent:
      type: object
      required:
      - id
      - argsMap
      properties:
        argsMap:
          type: object
        id:
          $ref: '#/components/schemas/ReferenceId'
    FileFormat:
      type: string
      enum:
      - png
      - svg
      - pdf
      - raw-svg
...
    FocusRootNodesInputEvent:
      default: null
      nullable: true
...
    FrameNode:
      type: object
      required:
      - size
      properties:
        blendMode:
          $ref: '#/components/schemas/BlendMode'
        children:
          type: array
          items:
            $ref: '#/components/schemas/ReferenceIdOrEntity'
        clipContent:
          type: boolean
        cornerRadii:
          $ref: '#/components/schemas/CornerRadii'
        id:
          allOf:
          - $ref: '#/components/schemas/ReferenceId'
          nullable: true
        layoutElement:
          $ref: '#/components/schemas/LayoutElement'
        layoutParent:
          allOf:
          - $ref: '#/components/schemas/StaticLayoutParent'
          nullable: true
        opacity:
          $ref: '#/components/schemas/Opacity'
        rotationDeg:
          $ref: '#/components/schemas/Angle'
        size:
          $ref: '#/components/schemas/Size'
        styles:
          type: array
          items:
            $ref: '#/components/schemas/Style'
        translation:
          $ref: '#/components/schemas/Vec2'
        visible:
          type: boolean
...
    LuaScriptArgsMap:
      type: object
      additionalProperties: {}
    LuaScriptWithId:
      type: object
      required:
      - id
      - source
      properties:
        id:
          $ref: '#/components/schemas/ReferenceId'
        source:
          type: array
          items:
            type: string
...
    MoveEntityInputEvent:
      type: object
      required:
      - id
      properties:
        dx:
          type: number
          format: float
          nullable: true
        dy:
          type: number
          format: float
          nullable: true
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
    Node:
      oneOf:
      - allOf:
        - $ref: '#/components/schemas/FrameNode'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - Frame
      - allOf:
        - $ref: '#/components/schemas/RectangleNode'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - Rectangle
      - allOf:
        - $ref: '#/components/schemas/EllipseNode'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - Ellipse
      - allOf:
        - $ref: '#/components/schemas/StarNode'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - Star
      - allOf:
        - $ref: '#/components/schemas/PolygonNode'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - Polygon
      - allOf:
        - $ref: '#/components/schemas/TextNode'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - Text
      - allOf:
        - $ref: '#/components/schemas/VectorNode'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - Vector
      discriminator:
        propertyName: type
...
    Paint:
      oneOf:
      - allOf:
        - $ref: '#/components/schemas/SolidPaint'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - Solid
      - allOf:
        - $ref: '#/components/schemas/ImagePaint'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - Image
      - allOf:
        - $ref: '#/components/schemas/GradientPaint'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - Gradient
      discriminator:
        propertyName: type
...
    RectangleNode:
      type: object
      required:
      - size
      properties:
        blendMode:
          $ref: '#/components/schemas/BlendMode'
        cornerRadii:
          $ref: '#/components/schemas/CornerRadii'
        id:
          allOf:
          - $ref: '#/components/schemas/ReferenceId'
          nullable: true
        layoutElement:
          $ref: '#/components/schemas/LayoutElement'
        opacity:
          $ref: '#/components/schemas/Opacity'
        rotationDeg:
          $ref: '#/components/schemas/Angle'
        size:
          $ref: '#/components/schemas/Size'
        styles:
          type: array
          items:
            $ref: '#/components/schemas/Style'
        translation:
          $ref: '#/components/schemas/Vec2'
        visible:
          type: boolean
    ReferenceId:
      type: string
...
    RegisterLuaScriptInputEvent:
      type: object
      required:
      - script
      properties:
        script:
          $ref: '#/components/schemas/LuaScriptWithId'
...
    Size:
      type: array
      items:
        type: number
        format: float
      description: ''
      maxItems: 2
      minItems: 2
...
    StarNode:
      type: object
      required:
      - size
      properties:
        blendMode:
          $ref: '#/components/schemas/BlendMode'
        id:
          allOf:
          - $ref: '#/components/schemas/ReferenceId'
          nullable: true
        innerRadiusRatio:
          type: number
          format: float
        layoutElement:
          $ref: '#/components/schemas/LayoutElement'
        opacity:
          $ref: '#/components/schemas/Opacity'
        pointCount:
          type: integer
          format: int32
          minimum: 0
        rotationDeg:
          $ref: '#/components/schemas/Angle'
        size:
          $ref: '#/components/schemas/Size'
        styles:
          type: array
          items:
            $ref: '#/components/schemas/Style'
        translation:
          $ref: '#/components/schemas/Vec2'
        visible:
          type: boolean
...
    UpdateCanvasSizeInputEvent:
      type: object
      required:
      - size
      properties:
        size:
          $ref: '#/components/schemas/Size'
    UpdateCanvasViewportInputEvent:
      type: object
      required:
      - viewport
      properties:
        viewport:
          $ref: '#/components/schemas/Viewport'
    UpdateDropShadowStyleInputEvent:
      type: object
      required:
      - id
      properties:
        blur:
          allOf:
          - $ref: '#/components/schemas/Abs'
          nullable: true
        color:
          allOf:
          - $ref: '#/components/schemas/Color'
          nullable: true
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        position:
          allOf:
          - $ref: '#/components/schemas/Vec2'
          nullable: true
        spread:
          allOf:
          - $ref: '#/components/schemas/Abs'
          nullable: true
    UpdateEllipseNodeInputEvent:
      type: object
      required:
      - id
      properties:
        endingAngle:
          type: number
          format: float
          nullable: true
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        innerRadiusRatio:
          type: number
          format: float
          nullable: true
        startingAngle:
          type: number
          format: float
          nullable: true
    UpdateEntityBlendModeInputEvent:
      type: object
      required:
      - id
      - blendMode
      properties:
        blendMode:
          $ref: '#/components/schemas/BlendMode'
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
    UpdateEntityChildrenInputEvent:
      type: object
      required:
      - id
      - children
      properties:
        children:
          type: array
          items:
            $ref: '#/components/schemas/ReferenceIdOrEntity'
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
    UpdateEntityCornerRadiiInputEvent:
      type: object
      required:
      - id
      - cornerRadii
      properties:
        cornerRadii:
          $ref: '#/components/schemas/CornerRadii'
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
    UpdateEntityOpacityInputEvent:
      type: object
      required:
      - id
      - opacity
      properties:
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        opacity:
          $ref: '#/components/schemas/Opacity'
    UpdateEntityRotationInputEvent:
      type: object
      required:
      - id
      - rotationDeg
      properties:
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        rotationDeg:
          $ref: '#/components/schemas/Angle'
    UpdateEntitySizeInputEvent:
      type: object
      required:
      - id
      - size
      properties:
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        size:
          $ref: '#/components/schemas/Size'
    UpdateEntityTransformInputEvent:
      type: object
      required:
      - id
      properties:
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        rotationDeg:
          allOf:
          - $ref: '#/components/schemas/Angle'
          nullable: true
        x:
          type: number
          format: float
          nullable: true
        y:
          type: number
          format: float
          nullable: true
    UpdateEntityVisibilityInputEvent:
      type: object
      required:
      - id
      - visible
      properties:
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        visible:
          type: boolean
    UpdateFillStyleInputEvent:
      type: object
      required:
      - id
      properties:
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        paintId:
          allOf:
          - $ref: '#/components/schemas/ReferenceIdOrEntity'
          nullable: true
    UpdateFrameNodeInputEvent:
      type: object
      required:
      - id
      properties:
        clipContent:
          type: boolean
          nullable: true
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
    UpdateGradientPaintInputEvent:
      type: object
      required:
      - id
      properties:
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        stops:
          type: array
          items:
            $ref: '#/components/schemas/GradientColorStop'
          nullable: true
        variant:
          allOf:
          - $ref: '#/components/schemas/GradientVariant'
          nullable: true
    UpdateImagePaintInputEvent:
      type: object
      required:
      - id
      properties:
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        imageId:
          allOf:
          - $ref: '#/components/schemas/ImageId'
          nullable: true
        scaleMode:
          allOf:
          - $ref: '#/components/schemas/ImageScaleMode'
          nullable: true
    UpdatePolygonNodeInputEvent:
      type: object
      required:
      - id
      properties:
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        pointCount:
          type: integer
          format: int32
          nullable: true
          minimum: 0
    UpdateSolidPaintInputEvent:
      type: object
      required:
      - id
      properties:
        color:
          $ref: '#/components/schemas/Color'
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
    UpdateStarNodeInputEvent:
      type: object
      required:
      - id
      properties:
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        innerRadiusRatio:
          type: number
          format: float
          nullable: true
        pointCount:
          type: integer
          format: int32
          nullable: true
          minimum: 0
    UpdateStorkeStyleInputEvent:
      type: object
      required:
      - id
      properties:
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        paintId:
          allOf:
          - $ref: '#/components/schemas/ReferenceIdOrEntity'
          nullable: true
        width:
          allOf:
          - $ref: '#/components/schemas/Abs'
          nullable: true
    UpdateTextNodeInputEvent:
      type: object
      required:
      - id
      properties:
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/TextAttributeInterval'
          nullable: true
        horizontalTextAlignment:
          allOf:
          - $ref: '#/components/schemas/HorizontalTextAlignment'
          nullable: true
        id:
          $ref: '#/components/schemas/ReferenceIdOrEntity'
        lineWrap:
          allOf:
          - $ref: '#/components/schemas/LineWrap'
          nullable: true
        sizingMode:
          allOf:
          - $ref: '#/components/schemas/TextSizingMode'
          nullable: true
        text:
          type: string
          nullable: true
        verticalTextAlignment:
          allOf:
          - $ref: '#/components/schemas/VerticalTextAlignment'
          nullable: true
...
    Viewport:
      type: object
      required:
      - physicalPosition
      - physicalSize
      properties:
        physicalPosition:
          $ref: '#/components/schemas/Vec2'
        physicalSize:
          $ref: '#/components/schemas/Size'
...